Update base image to 9.0.4
This switches the base image to ubuntu-base:9.0.4, ie ubuntu 22.04.
The base deps versions need to be updated.
libcap2 is already installed in the base image, so it's removed from the
list of packages to install

Only the 6.0 and 7.0 repositories contain 'mongodb-org-server' packages
for ubuntu 22.04, so we are forced to use one of these versions.
However, the unifi package requires a mongodb-org-server version <<
5.0.0 so we need to ignore the mongodb-org-server dependency when
installing the unifi package.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>